Spindle improvements (wip)
- extracted table

#3256
