+++ waltz-ng/client/measurable-relationship/components/panel/related-measurables-panel.html
@@ -37,51 +37,12 @@
         <div ng-if="$ctrl.gridData.length > 0"
              class="waltz-fade-if small">
             <div ng-class="{ 'waltz-scroll-region-500' : $ctrl.gridData.length > 10 }">
-                <table class="table table-condensed table-hover">
-                    <colgroup>
-                        <col width="38%">
-                        <col width="20%">
-                        <col width="38%">
-                        <col width="4%">
-                    </colgroup>
-                    <thead>
-                    <tr>
-                        <th>From</th>
-                        <th>Relationship</th>
-                        <th>To</th>
-                        <th>..</th>
-                    </tr>
-                    </thead>
-                    <tbody>
-                    <!-- DATA -->
-                    <tr ng-repeat="row in $ctrl.gridData"
-                        class="clickable"
-                        ng-class="{ 'waltz-highlighted-row' : $ctrl.isSelected(row) }"
-                        ng-click="$ctrl.selectRow(row)">
-                        <td>
-                            <span ng-class="{ 'text-muted' : row.outbound }">
-                                <span ng-bind="row.a.name"></span>
-                                ( <small class='text-muted' ng-bind="row.a.type"></small> )
-                            </span>
-                        </td>
-                        <td>
-                            <span ng-bind="row.relationship.relationship | toDisplayName:'relationshipKind'"></span>
-                        </td>
-                        <td>
-                            <span ng-class="{ 'text-muted' : !row.outbound }">
-                                <span ng-bind="row.b.name"></span>
-                                ( <small class='text-muted' ng-bind="row.b.type"></small> )
-                            </span>
-                        </td>
-                        <td>
-                            <span ng-if="row.relationship.description.length > 0"
-                                  title="Additional description available">
-                                <waltz-icon name="file-text-o"></waltz-icon>
-                            </span>
-                        </td>
-                    </tr>
-                    </tbody>
-                </table>
+
+                <waltz-related-measurables-table rows="$ctrl.gridData"
+                                                 on-row-select="$ctrl.selectRow"
+                                                 selected-row="$ctrl.selectedRow">
+
+                </waltz-related-measurables-table>
             </div>
 
             <!-- Export -->
