+++ waltz-ng/client/measurable-relationship/components/panel/related-measurables-panel.js
@@ -219,15 +219,6 @@ function controller($q, $timeout, serviceBroker, notification) {
         vm.selectedRow = null;
     };
 
-    vm.isSelected = (row) => {
-        if (vm.selectedRow) {
-            const sameA = sameRef(row.a, vm.selectedRow.a);
-            const sameB = sameRef(row.b, vm.selectedRow.b);
-            return sameA && sameB;
-        } else {
-            return false;
-        }
-    };
 
     vm.removeRelationship = (rel) => {
         if (confirm('Are you sure you want to delete this relationship ?')) {
