+++ waltz-ng/client/measurable-relationship/components/table/related-measurables-table.js
@@ -1,0 +1,63 @@
+/*
+ * Waltz - Enterprise Architecture
+ * Copyright (C) 2016, 2017 Waltz open source project
+ * See README.md for more information
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import template from './related-measurables-table.html';
+import {initialiseData} from "../../../common/index";
+import {sameRef} from "../../../common/entity-utils";
+
+
+const bindings = {
+    rows: '<',
+    onRowSelect: '<',
+    selectedRow: '<'
+};
+
+
+const initialState = {
+    rows: [],
+    onRowSelect: (row) => console.log('default on row select', { row })
+};
+
+
+function controller() {
+    const vm = initialiseData(this, initialState);
+
+    vm.isSelected = (row) => {
+        if (vm.selectedRow) {
+            const sameA = sameRef(row.a, vm.selectedRow.a);
+            const sameB = sameRef(row.b, vm.selectedRow.b);
+            return sameA && sameB;
+        } else {
+            return false;
+        }
+    };
+}
+
+
+const component = {
+    template,
+    controller,
+    bindings
+};
+
+
+export default {
+    component,
+    id: 'waltzRelatedMeasurablesTable'
+}
